requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pycryptodome==3.20.0
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
webargs==8.4.0
Werkzeug==3.0.3
wrapt==1.16.0